refactor: fix code convention
